Refactor main window into GUI and logic modules
