+++ dialogs/__init__.py
@@ -1,6 +1,5 @@
 from .batch_mapping_dialog import BatchMappingDialog
 from .mapping_dialog import MappingDialog
-from .drop_area import DropArea
 from .download_dialog import DownloadDialog
 
-__all__ = ["BatchMappingDialog", "MappingDialog", "DropArea", "DownloadDialog"]
+__all__ = ["BatchMappingDialog", "MappingDialog", "DownloadDialog"]
