--- processor.py
@@ -1,5 +1,0 @@
-"""Реэкспорт класса ExcelToGoogleSheets для обратной совместимости."""
-
-from business.processor import ExcelToGoogleSheets
-
-__all__ = ["ExcelToGoogleSheets"]
